Fix puli/issues#178
